Tooltip

+++ Gwen/Input/KeyData.cs
@@ -9,7 +9,7 @@ namespace Gwen.Input
     public class KeyData
     {
         public readonly bool[] KeyState;
-        public readonly float [] NextRepeat;
+        public readonly double [] NextRepeat;
         public Controls.ControlBase Target;
         public bool LeftMouseDown;
         public bool RightMouseDown;
@@ -17,7 +17,7 @@ namespace Gwen.Input
         public KeyData()
         {
             KeyState = new bool[(int)Key.Count];
-            NextRepeat = new float[(int)Key.Count];
+            NextRepeat = new double[(int)Key.Count];
             // everything is initialized to 0 by default
         }
     }
